Temporary commit: many classes shall be removed because are old

+++ src/net/KNet/Developed/Org/Apache/Kafka/Clients/Producer/ProducerRecord.cs
@@ -32,7 +32,7 @@ namespace Org.Apache.Kafka.Clients.Producer
         {
         }
 
-        public System.DateTime DateTime => System.DateTimeOffset.FromUnixTimeMilliseconds(Timestamp).DateTime;
+        public System.DateTime DateTime => System.DateTimeOffset.FromUnixTimeMilliseconds((long)Timestamp()).DateTime;
     }
 }
