+++ src/net/KNet/Developed/Org/Apache/Kafka/Clients/Producer/RecordMetadata.cs
@@ -20,6 +20,6 @@ namespace Org.Apache.Kafka.Clients.Producer
 {
     public partial class RecordMetadata
     {
-        public System.DateTime DateTime => System.DateTimeOffset.FromUnixTimeMilliseconds(Timestamp).DateTime;
+        public System.DateTime DateTime => System.DateTimeOffset.FromUnixTimeMilliseconds((long)Timestamp()).DateTime;
     }
 }
